refactor(server): renaming Language to ProgramingLanguage
